Refactor user service and rearrange elements in feed
